requirements.txt
Django>=3.0.0,<3.1.0
djangorestframework>=3.11.0,<3.12.0
psycopg2>=2.8.5,<2.9.0
django-cors-headers>=3.5.0,<3.6.0

flake8>=3.8.0,<3.9.0